+++ Trains/OpenFreightCar.h
@@ -29,7 +29,7 @@ public:
     virtual string getTypeString() const { return "Open Freight Car"; }
 
     // -------- SETTERS --------
-    virtual void setArea(double area) { _area = area; } // Area in square meters
+    virtual void setArea(int area) { _area = area; } // Area in square meters
 };
 
 #endif // !OPENFREIGHTCAR_H
